+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Fixed
+* Add back github icon for old site passivity
 
 6.0.0 - (July 30, 2019)
 ----------