+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Unreleased
 
 ### changed
 * Remove div containing status view to center component.
+* Raw routes are wrapped in main.
 
 6.0.0 - (July 30, 2019)
 ----------