+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+6.1.0 - (August 13, 2019)
+----------
 ### Fixed
 * Add back github icon for old site passivity.
 * Scroll to location if an anchor is present.